requirements.txt
discord.py[voice]==1.3.1
xkcd==2.4.2
requests==2.23.0
sqlitedict==1.6.0
